+++ app/src/main/java/com/example/emptyactivity/LoginPage.kt
@@ -1,6 +1,5 @@
-package com.example.emptyactivity.components
+package com.example.emptyactivity
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.height
@@ -18,16 +17,16 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 
-
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun SignupPage(modifier: Modifier = Modifier) {
+fun LoginPage(modifier: Modifier = Modifier) {
     Column(
         modifier = modifier,
         horizontalAlignment = Alignment.CenterHorizontally,
         verticalArrangement = Arrangement.Center,
     ) {
         Text(
-            text = "Create an Account",
+            text = "Log In",
             style = MaterialTheme.typography.headlineMedium,
             fontWeight = FontWeight.Bold,
         )
@@ -39,10 +38,6 @@ fun SignupPage(modifier: Modifier = Modifier) {
             value = "Password",
             onValueChange = {},
         )
-        LoginTextField(
-            value = "Repeat Password",
-            onValueChange = {},
-        )
         Button(
             modifier = Modifier
                 .padding(10.dp)
@@ -51,24 +46,25 @@ fun SignupPage(modifier: Modifier = Modifier) {
             onClick = {}
         ) {
             Text(
-                text = "Sign Up",
+                text = "Log in",
                 style = MaterialTheme.typography.titleLarge
-            )
-        }
-
-        Button(
-            onClick = { /*TODO*/ },
-            modifier = Modifier.background(MaterialTheme.colorScheme.background)
-        ) {
-            Text(
-                text = "Don't have an account? Sign up"
             )
         }
     }
 }
 
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun LoginTextField(value: String, onValueChange: (String) -> Unit, modifier: Modifier = Modifier) {
+    TextField(
+        value = value,
+        onValueChange = onValueChange,
+        modifier = modifier.padding(5.dp)
+    )
+}
+
 @Preview
 @Composable
-fun SignupPagePreview(modifier: Modifier = Modifier) {
-    SignupPage(modifier)
+fun LoginPagePreview(modifier: Modifier = Modifier) {
+    LoginPage(modifier)
 }
